@@ -71,7 +71,7 @@ PACKAGE_MAPPINGS = {
     'sentence_transformers': 'sentence-transformers',
     'rank_bm25': 'rank-bm25',
     'snowballstemmer': 'snowballstemmer',
-    'PyPDF2': 'PyPDF2',
+    'pypdf': 'pypdf',
     'pdf2image': 'pdf2image',
 }
 
